Add files via upload
There are ‘n’ number of villages. You are given an array of size ‘n’ representing the population of each village. Every year, there is a cricket competition between two teams and villagers come to see the match. Villagers from ith village and (n-i)-1th village (0 <= i < n/2) are combined and then formed groups of 10 people each. For e.g. villagers from village 0 and n-1, 1 and n-2, 2 and n-3 are combined. The number of villages are always even. So, clearly there will be n/2 combinations from all the villages. You have to tell that how many groups will be formed in each combination and how many villagers will be left without the complete group of 10 peoples.